Define GoogleStrategy Callback function
- I have categorized two cases: those who have signed up before and
those who have never signed up.
- After updating the logined data in Database, done method is going to
operate.
- Done is a method called internally by the strategy implementation. It
navigates you, one of the success/error/fail methods. Each of these
options may calls to the next.When success is called, it can attach the
user to the request or do other things depending on your needs, and send
it to serialize function that I'm going to define