Understanding Passport Serialize and Deserialize
- After done method in googleStrategy callback called, user.id is sent
to passport.serialize and passport.serialize is going to determines
which data of the user object should be stored in the session. The
result of the serializeUser method is attached to the session as
req.session.passport.user = {}. For example, req.session.passport.user
= {id: 'abc'}.

- Then, the first argument of deserialize corresponds to the key of the
user object that was given to the done function. So your whole object is
retrieved with help of that key. That key here is the user id (key can
be any key of the user Object. Finally, done method in deserializeUser
will attach this object to the request as req.user